+++ src/test/java/com/poorlex/poorlex/user/member/service/MemberQueryServiceTest.java
@@ -12,7 +12,6 @@ import com.poorlex.poorlex.battlesuccess.domain.BattleSuccessHistoryRepository;
 import com.poorlex.poorlex.consumption.expenditure.domain.Expenditure;
 import com.poorlex.poorlex.consumption.expenditure.domain.ExpenditureRepository;
 import com.poorlex.poorlex.consumption.expenditure.fixture.ExpenditureFixture;
-import com.poorlex.poorlex.consumption.expenditure.service.dto.response.MyPageExpenditureResponse;
 import com.poorlex.poorlex.fixture.MemberFixture;
 import com.poorlex.poorlex.friend.domain.Friend;
 import com.poorlex.poorlex.friend.domain.FriendRepository;
@@ -26,6 +25,7 @@ import com.poorlex.poorlex.user.member.domain.MemberLevel;
 import com.poorlex.poorlex.user.member.domain.MemberRepository;
 import static com.poorlex.poorlex.user.member.domain.Oauth2RegistrationId.APPLE;
 import com.poorlex.poorlex.user.member.service.dto.ExpenditureDto;
+import com.poorlex.poorlex.user.member.service.dto.response.MyPageExpenditureResponse;
 import com.poorlex.poorlex.user.member.service.dto.response.MyPageResponse;
 import com.poorlex.poorlex.user.point.domain.MemberPoint;
 import com.poorlex.poorlex.user.point.domain.MemberPointRepository;
